refactor: convert Price back to stringly typed
Use convenience wrappers in code instead, pricing terms are so dynamic it is difficult to create a static system for it, plus it makes storing it much more difficult than it should. The static system does not help too much as the application is only concerned about presenting it as a string rather than process it.