feat: use database as datasource
Store places in a json file, allowing for easy migration to server in the future. We ideally want to use the database instead of an in-memory app to avoid large memory consumption due to many entries. We also get Full-Text-Search for free.

Use a prepackaged database for simplicity, a sql scripted is attached to create the database. In the future the application should receive data from the server and store responses in Room for offline first functionality.

Start moving images to urls, in the future use Coil to load images. Bundling Images as part of the APK is a bad idea.